cleanup python
